+++ syntax/specialization/src/main/scala/org/enso/syntax/text/Parser.scala
@@ -2,12 +2,11 @@ package org.enso.syntax.text
 
 import org.enso.flexer.Macro.compile
 import org.enso.flexer.Success
-import org.enso.syntax.text.ast.Renamer
-import org.enso.syntax.text.parser.Definition
+import org.enso.syntax.text.precedence.Mixfix
 import org.enso.flexer
+import org.enso.syntax.text
 import org.enso.syntax.text.AST
-
-import org.enso.syntax.text.ast.Renamer.partition
+import org.enso.syntax.text.precedence.Mixfix.partition
 
 ////////////////
 //// Parser ////
@@ -19,21 +18,14 @@ class Parser {
 
   def run(input: String): Result[AST.Module] = engine.run(input).map { module =>
     val module2 = module.asInstanceOf[AST.Module] // FIXME
-//    val ppp     = partition(module2.firstLine.elem.get)
-//    println("\n========= 1\n")
-//    pprint.pprintln(ppp, width = 50, height = 1000)
-//    val tt: SpacedList[AST] = SpacedList(ppp.head.el, ppp.tail)
-//    println("\n========= 2\n")
-//    pprint.pprintln(Renamer.run(tt), width = 50, height = 1000)
-    Renamer.run(module2)
-//    module2
+    Mixfix.run(module2)
   }
 
 }
 
 object Parser {
   type Result[T] = flexer.Result[T]
-  private val newEngine = compile(parser.Definition)
+  private val newEngine = compile(text.ParserDef)
 }
 
 //////////////
@@ -41,39 +33,6 @@ object Parser {
 //////////////
 
 object Main extends App {
-
-  var indent = 0
-
-//  def pprint[T](t: T) {
-//    if (t == null) {
-//      println(t)
-//      return
-//    }
-//    val s = t.toString()
-//    print("  " * indent)
-//    val (l, r2) = s.span(x => x != '(' && x != ')')
-//    print(l)
-//    if (r2 == "") {
-//      println
-//      return
-//    }
-//
-//    val (m, r) = r2.splitAt(1)
-//
-//    if (m == "(") {
-//      indent += 1
-//      println(m)
-//      pprint(r)
-//    } else if (m == ")") {
-//      indent -= 1
-//      println(m)
-//      pprint(r)
-//    }
-//
-//  }
-
-  val parserCons = compile(Definition)
-
   val p1 = new Parser()
   val p2 = new Parser()
